Fixed bottom layer
Reoriented the model to print flat. 
Also, the added cable clip was 0.4mm below the rest of the part in Cura 5.2.1. I imported the model to Tinkercad and exported, this now works in Cura.
Also, it appears this is the only modded part. The other files in this folder are Voron stock.